Refactor room components and integrate new features for improved functionality
- Updated import paths for RoomList component to reflect new directory structure.
- Introduced new Chat, GameStatus, IntegratedQuizGame, PlayerList, and QuizManagement components to enhance room functionality.
- Refactored Room component to utilize new components, improving code organization and readability.
- Added comprehensive tests for the RoomList component to ensure proper functionality and user experience.
- Enhanced socket communication for real-time updates and interactions within the room.
